chore(ts): phase 4c - migrate heavy system hooks
Migrate the two larger system hooks that are more involved than the
phase 4a barrels but don't pull in adapters/media/audio yet.

- useDeepLink: typed UseDeepLinkOptions, DeepLinkStateSnapshot for the
  ref-based closure-avoidance pattern, RobotStatus/BusyReason from
  types/robot, ToastSeverity from types/store. Removed the now-needless
  eslint-disable since exhaustive-deps is happy with [] here.
- usePermissions: UsePermissionsResult, PermissionSnapshot (null-first
  booleans), OptionalBoolResult alias for the Rust Option<bool>
  commands (check_local_network_permission, check_location_permission,
  check_bluetooth_permission). Added a minimal global augmentation for
  window.__E2E_MODE__.

Verified: typecheck, 55 unit tests, lint (no new warnings on touched
files), production build. No behavioural change.

Identified latent issues (not introduced here, tracked in TODO):
- useDeepLink: cleanup race if unmount happens before setupListener
  resolves (inoffensive: hook lives for app lifetime).
- usePermissions: isE2EMode() runs on every render and logs on each
  detection branch (noisy in E2E runs).

Made-with: Cursor
